add a hack to start the image with the remembered window size again
This does not feel like the ultimate solution but I have not found a
better place to restore these dimensions. It needs an Interpreter to do
that (for the reference to the image, space has none), yet one is not
usually available when dealing with the display in FormWrapper or
W_DisplayBitmap.