Bugfixes and improvements
Rename model/DBObject.py to model/database.py for
consistency.
Fix a bug in temporary inherits_from() function
which causes segfaults when loading code.
Replace db_mode with DBManager().driver because
we are already maintaining an instance of
DBManager. This also allows to reference a driver
directly rather than having to deal with
constants. The driver selection only happens in
primitiveSQLModeSwitch now.