more coarse-grained caches for compiled methods.
CompiledMethods will be made immutable in the image, but until then even
now to make changes to compiled methods valid users need to call
flushCache either on the method or the class objects anyway. We're now
marking many methods in the compiled methods simply elidable, and rely
on prudent use of flushCache in the image. all the flushCache primitives
go to the class and invalidate their method lookup caches, so this will
invalidate any traces that include sends to the changed method. The only
care has to be taken for become and becomeForward, where we just blow
the caches for all involved classes.